requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pytest
azure-storage-blob
python-dotenv
